requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
black==25.9.0
cffi==2.0.0
click==8.3.0
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.115.0
h11==0.16.0
httptools==0.7.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
mypy_extensions==1.1.0
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.5.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
python-jose==3.5.0
pytokens==0.2.0
PyYAML==6.0.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.34
starlette==0.38.6
typing_extensions==4.15.0
uvicorn==0.30.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
pydantic[email]
